requirements.txt
cryptography==39.0.0
pycryptodome==3.17.0
zlib-wrapper==0.1.3
flask==2.2.3
Werkzeug==2.2.3
requests==2.28.2 